notepad_tracker/requirements.txt
Flask>=2.0
GitPython>=3.1
python-dotenv>=0.21
